refactor(hooks/builtins): let DMR own its unload conventions
Address review feedback on #2706:

- Export pkg/model/provider/dmr.ProviderType and UnloadURL so the
  unload builtin becomes a dumb dispatcher and DMR owns the provider
  literal + the /v1 -> /_unload URL convention (sibling to the
  existing /_configure helper). Move the URL-resolution table test
  into the dmr package where the helper now lives.
- Gate the FromAgentModels snapshot in executeOnAgentSwitchHooks on
  executor.Has(EventOnAgentSwitch) so audit-free deployments don't
  pay the team-lookup + per-model allocation on every agent switch.
- Add a mixed-providers test pinning the per-element DMR filter:
  cloud entries (openai, anthropic) in the snapshot must be silently
  skipped, only DMR endpoints get POSTed.
- Add a comment on http.DefaultClient explaining why the SSRF-safe
  client used by http_post is wrong here (DMR is loopback).
